refactor: move DrawScope extension function into new package
